# core/generated/swagger-ui/schema-registry-api-spec.yaml
...
swagger: "2.0"
info:
  version: "v1"
  title: "Confluent Schema Registry"
schemes:
- "http"
- "https"
paths:
...
  /compatibility/subjects/{subject}/versions/{version}:
    post:
      summary: "Test input schema against a particular version of a subject's schema\
        \ for compatibility."
      description: "the compatibility level applied for the check is the configured\
        \ compatibility level for the subject (http:get:: /config/(string: subject)).\
        \ If this subject's compatibility level was never changed, then the global\
        \ compatibility level applies (http:get:: /config)."
      operationId: "testCompatabilityBySubjectName"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "Content-Type"
        in: "header"
        required: false
        type: "string"
      - name: "Accept"
        in: "header"
        required: false
        type: "string"
      - name: "subject"
        in: "path"
        description: "Subject of the schema version against which compatibility is\
          \ to be tested"
        required: true
        type: "string"
      - name: "version"
        in: "path"
        description: "Version of the subject's schema against which compatibility\
          \ is to be tested. Valid values for versionId are between [1,2^31-1] or\
          \ the string \"latest\".\"latest\" checks compatibility of the input schema\
          \ with the last registered schema under the specified subject"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Schema"
        required: true
        schema:
          $ref: "#/definitions/RegisterSchemaRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/CompatibilityCheckResponse"
        404:
          description: "Error code 40401 -- Subject not found\nError code 40402 --\
            \ Version not found"
        422:
          description: "Error code 42201 -- Invalid Avro schema\nError code 42202\
            \ -- Invalid version"
        500:
          description: "Error code 50001 -- Error in the backend data store"
...
  /schemas/ids/{id}:
    get:
      summary: "Get the schema string identified by the input ID."
      description: ""
      operationId: "getSchema"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "id"
        in: "path"
        description: "Globally unique identifier of the schema"
        required: true
        type: "integer"
        format: "int32"
      - name: "fetchMaxId"
        in: "query"
        required: false
        type: "boolean"
        default: false
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/SchemaString"
        404:
          description: "Error code 40403 -- Schema not found\n"
        500:
          description: "Error code 50001 -- Error in the backend data store\n"
...
  /subjects/{subject}/versions/{version}:
    get:
      summary: "Get a specific version of the schema registered under this subject."
      description: ""
      operationId: "getSchemaByVersion"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "subject"
        in: "path"
        description: "Name of the Subject"
        required: true
        type: "string"
      - name: "version"
        in: "path"
        description: "Version of the schema to be returned. Valid values for versionId\
          \ are between [1,2^31-1] or the string \"latest\". \"latest\" returns the\
          \ last registered schema under the specified subject. Note that there may\
          \ be a new latest schema that gets registered right after this request is\
          \ served."
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Schema"
        404:
          description: "Error code 40401 -- Subject not found\nError code 40402 --\
            \ Version not found"
        422:
          description: "Error code 42202 -- Invalid version"
        500:
          description: "Error code 50001 -- Error in the backend data store"
    delete:
      summary: "Deletes a specific version of the schema registered under this subject.\
        \ This only deletes the version and the schema ID remains intact making it\
        \ still possible to decode data using the schema ID. This API is recommended\
        \ to be used only in development environments or under extreme circumstances\
        \ where-in, its required to delete a previously registered schema for compatibility\
        \ purposes or re-register previously registered schema."
      description: ""
      operationId: "deleteSchemaVersion"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "subject"
        in: "path"
        description: "Name of the Subject"
        required: true
        type: "string"
      - name: "version"
        in: "path"
        description: "Version of the schema to be returned. Valid values for versionId\
          \ are between [1,2^31-1] or the string \"latest\". \"latest\" returns the\
          \ last registered schema under the specified subject. Note that there may\
          \ be a new latest schema that gets registered right after this request is\
          \ served."
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "integer"
            format: "int32"
        404:
          description: "Error code 40401 -- Subject not found\nError code 40402 --\
            \ Version not found"
        422:
          description: "Error code 42202 -- Invalid version"
        500:
          description: "Error code 50001 -- Error in the backend data store"
...
definitions:
  CompatibilityCheckResponse:
    type: "object"
    properties:
      is_compatible:
        type: "boolean"
...
  RegisterSchemaRequest:
    type: "object"
    properties:
      version:
        type: "integer"
        format: "int32"
      id:
        type: "integer"
        format: "int32"
      schemaType:
        type: "string"
      schema:
        type: "string"
...
  Schema:
    type: "object"
    properties:
      subject:
        type: "string"
      version:
        type: "integer"
        format: "int32"
      id:
        type: "integer"
        format: "int32"
      schema:
        type: "string"
      schemaType:
        type: "string"
  SchemaString:
    type: "object"
    properties:
      schema:
        type: "string"
        description: "Schema string identified by the ID"
      maxId:
        type: "integer"
        format: "int32"
...